Fix Bug with rotating joints
Joints now easly, and smoothly rotate!
Before it was all fucked up.
You moved one joint but it also eneded up
moving another joint. It was a buggy hell!

Also, I did the one thing you not supposed to,
my first thought was to blame it on the 3rd
party software! What an idiot, its always
my (the developers) fault!